requirements.txt
streamlit
python-dotenv
PyPDF2
langchain-core
langchain-community
transformers
torch
sarvamai
sentence-transformers
faiss-cpu
langchain_google_genai